Move EventBuilder to event.go, reduce allocs when parsing events
